+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog for bulletproofs
 
+## 2.0
+
+* Use elliptic-curve library as dependency
+* Update to galois-field-1.0
+
 ## 1.0.1
 
 * Fix arithmoi dependency.